+++ src/core/controllers/index.ts
@@ -2,4 +2,4 @@ export * from "./FPSController.js";
 export * from "./OrbitController.js";
 export * from "./WASDController.js";
 export * from "./ZoomController.js";
-export * from "./DeviceOrientationController.js";
+export * from "./FlyController.js";